# openapi/containers-api.yaml
openapi: 3.0.3
info:
  title: Andy Containers API
  description: |
    Development container management platform for the Andy ecosystem.
    Manages containers, workspaces, templates, infrastructure providers, and container images
    across heterogeneous infrastructure (Docker, Apple Containers, Azure, SSH, Rivoli-managed).
  version: 0.1.0-alpha
  contact:
    name: Rivoli AI
    url: https://github.com/rivoli-ai/andy-containers
servers:
  - url: https://containers.rivoli.ai
    description: Production
  - url: https://containers-uat.rivoli.ai
    description: UAT
  - url: https://localhost:5200
    description: Local Development
security:
  - bearerAuth: []
paths:
...
  /api/workspaces:
    get:
      tags: [Workspaces]
      summary: List workspaces
      parameters:
        - name: ownerId
          in: query
          schema: { type: string }
        - name: organizationId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { type: string, enum: [Active, Suspended, Archived] }
        - name: skip
          in: query
          schema: { type: integer, default: 0 }
        - name: take
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: List of workspaces
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Workspace' }
                  totalCount: { type: integer }
    post:
      tags: [Workspaces]
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWorkspaceRequest' }
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
  /api/workspaces/{id}:
    get:
      tags: [Workspaces]
      summary: Get workspace details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Workspace details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
    put:
      tags: [Workspaces]
      summary: Update a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateWorkspaceRequest' }
      responses:
        '200':
          description: Workspace updated
    delete:
      tags: [Workspaces]
      summary: Delete a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Workspace deleted

  # --- Templates (Catalog) ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Workspace:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        ownerId: { type: string }
        status: { type: string, enum: [Active, Suspended, Archived] }
        defaultContainerId: { type: string, format: uuid }
        gitRepositoryUrl: { type: string }
        gitBranch: { type: string }
        createdAt: { type: string, format: date-time }
        lastAccessedAt: { type: string, format: date-time }
    CreateWorkspaceRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        organizationId: { type: string, format: uuid }
        teamId: { type: string, format: uuid }
        gitRepositoryUrl: { type: string }
        gitBranch: { type: string }
        templateId: { type: string, format: uuid }
        templateCode: { type: string }
    UpdateWorkspaceRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        gitBranch: { type: string }
...